// types.ts
export enum View {
  DASHBOARD = 'DASHBOARD',
  COMPANIES = 'COMPANIES',
  VISITS = 'VISITS',
  TRAININGS = 'TRAININGS',
  ALERTS = 'ALERTS',
  AUDIT = 'AUDIT',
  INSPECTION_FORM = 'INSPECTION_FORM',
  EVIDENCES = 'EVIDENCES',
  CHECKLISTS = 'CHECKLISTS',
  RISKS = 'RISKS',
  REPORTS = 'REPORTS',
  EPI_MANAGEMENT = 'EPI_MANAGEMENT',
  LEGAL_DOCUMENTS = 'legal_documents',
  INCIDENT_MANAGEMENT = 'incident_management',
  NOTIFICATIONS = 'notifications',
  ADMIN = 'admin',
  SETTINGS = 'SETTINGS',
  HELP = 'HELP',
  SUBSCRIPTION = 'SUBSCRIPTION',
  ASO_MANAGEMENT = 'ASO_MANAGEMENT'
}